# openapi.yml
openapi: 3.0.0
info:
  description:
    "This is a sample server Twitter Clone server.  You can find out more
    about     Swagger at [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/)."
  version: 1.0.0
  title: Swagger Twitter Clone
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8082/v1
...
paths:
...
  /accounts:
    post:
      tags:
        - accounts
      summary: Creating an account
      description: ""
      operationId: addAcount
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: john
                  description: The username of the account
                password:
                  type: string
                  example: P@ssw0rd
                  description: Password of user
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
  "/accounts/{username}":
    get:
      tags:
        - accounts
      summary: Fetching an account
      description: ""
      operationId: findAccountByID
      parameters:
        - name: username
          in: path
          description: Username of account to return
          required: true
          example: john
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
...
components:
  securitySchemes:
    Auth:
      type: apiKey
      name: Authentication
      in: header
  schemas:
    Account:
      type: object
      properties:
        username:
          type: string
          description: The username of the account
          example: john
        display_name:
          type: string
          description: The account's display name
          example: ジョン
        create_at:
          type: string
          format: date-time
          description: The time the account was created
        followers_count:
          type: integer
          description: The number of followers for the account
          example: 52
        following_count:
          type: integer
          description: The number of accounts the given account is following
          example: 128
        note:
          type: string
          description: Biography of user
        avatar:
          type: string
          description: URL to the avatar image
        header:
          type: string
          description: URL to the header image
...